+++ cache_scripts/argparser.py
@@ -1,6 +1,11 @@
 from argparse import ArgumentParser, BooleanOptionalAction
 from typing import List
 
+from datetime import datetime
+import pathlib
+
+import config
+
 class CacheScriptsArgParser(ArgumentParser):
     def __init__(self, available_platforms: List[str], available_networks: List[str]):
         super().__init__(description="Main script to populate dao-analyzer cache")
@@ -29,6 +34,11 @@ class CacheScriptsArgParser(ArgumentParser):
             help="Removes the cache before updating"
         )
         self.add_argument(
+            "-F", "--delete-force",
+            action="store_true", default=False,
+            help="Removes the datawarehouse folder before doing anything"
+        )
+        self.add_argument(
             "--skip-daohaus-names",
             action="store_true", default=False,
             help="Skips the step of getting Daohaus Moloch's names, which takes some time"
@@ -46,4 +56,17 @@ class CacheScriptsArgParser(ArgumentParser):
             nargs="+",
             required=False,
             help="Collectors to run. For example: aragon/casts"
+        )
+        self.add_argument(
+            "--block-datetime",
+            required=False,
+            type=datetime.fromisoformat,
+            help="Get data up to a block datetime (input in ISO format)"
+        )
+        self.add_argument(
+            "-D", "--datawarehouse",
+            help="Specifies the destination folder of the datawarehouse",
+            required=False,
+            type=pathlib.Path,
+            default=config.DEFAULT_DATAWAREHOUSE
         )